refactor(huddle): restore desktop file-size ratchet by extracting grown modules
The pre-push ratchet caps huddle sources at 1000 lines; the latency
levers pushed tts.rs to 1072 and mod.rs to 1001. Extract along the
existing module seams instead of trimming comments:

- tts_streaming.rs: the env-gated streaming synthesis path
  (BUZZ_TTS_STREAMING / BUZZ_TTS_EMIT_FRAMES) moves out of the worker
  loop into synthesize_streaming(), same behavior, single call site.
- handle_cancel_or_shutdown + lock_player_ops move to
  tts_voice_transition.rs beside the cancellation state they operate
  on (speaker_cancellation already used them through that module).
- add_agent_to_huddle moves from mod.rs to commands.rs with the other
  huddle-mutating tauri commands; mod.rs keeps the re-export so
  lib.rs's import surface is unchanged.

No functional change. tts.rs 1072 -> 934, mod.rs 1001 -> 920, ratchet
green.
